+++ wasmaudioworklet/index.html
@@ -11,6 +11,7 @@
     </head>
     <body>
         <app-javascriptmusic></app-javascriptmusic>
+        <script src="https://cdn.jsdelivr.net/npm/assemblyscript@0.27.14/dist/web.js"></script>
         <script src="app.js" type="module"></script>
     </body>
 </html>
